Feature visualizer (#20)
* Working version

* Added documentation to readme

* Empty implementation of analyzer suspend

* Suspend functionality and refactoring

* refactor

* some rework

* several refactoring items

* API complete

* Code complete

*  more tracing

* More code

* Found crash

* Fixed crash

* Working library

* +

* Fixed spectrum basics
Control with Draw event is now CustomVisualizer
Sample showcases VU meter and spectrometer

* Added LED VU meter control with basic functionality

* Refactored into new structure and fixed dependencies

* Fixed the sample app

* More cleanup

* Fixed release mode crash

* Fixed blanked display when missing data frames

* Fixing breaking build

* Some refactoring to avoid breaking changes down the line.

* couple small tweeks

* Fixed high bitrate issue and added audiobuffer tests

* Fixed issues when building for ARM

* Small changes

* Merge Feature-CIStep 1 (#13)

* make buildscript build binaries and nuget

* versioning

* rename

* fix versioning

* publish the nupkg

* Revert "publish the nupkg"

This reverts commit 948d9ca3fe9eeddab8469294bfe1d512781b26bf.

* cleanup nuspec

* attempt to fix certs

* another attempt to fix build

* Add a CIBUILD configuration that exludes the test app

* disable test again

* Another attempt

* fix nuspec

* Fixed issue #19

* Added visualization to sample app

+++ source/lib/AudioAnalyzer.h
@@ -56,7 +56,7 @@ namespace AudioVisualizer
 			void Flush();
 
 			long GetPosition() const { return _spInputBuffer->GetPosition(); }
-			UINT32 GetDownsampleRate() const { return _spInputBuffer->GetDownsampleFactor(); }
+			UINT32 GetDownsampleRate() const { return (UINT32) _spInputBuffer->GetDownsampleFactor(); }
 		};
 	}
 }
